pyproject.toml
[tool.poetry]
name = "aggregator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.5.3"
SQLAlchemy = "^1.4.23"
psycopg2 = "^2.9.1"
requests = "^2.26.0"
Flask = "^2.0.1"
alembic = "^1.6.5"
beautifulsoup4 = "^4.9.3"
lxml = "^4.6.3"
gunicorn = "^20.1.0"
python-slugify = "^5.0.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
